refactor: merge element and textnode adapters into dom module.
