+++ webf/lib/dom.dart
@@ -4,6 +4,7 @@
  */
 
 export 'src/dom/element.dart';
+export 'src/dom/element_widget_adapter.dart';
 export 'src/dom/event.dart';
 export 'src/dom/event_target.dart';
 export 'src/dom/character_data.dart';
