+++ webf/lib/widget.dart
@@ -4,14 +4,7 @@
  */
 
 export 'src/widget/webf.dart';
-export 'src/widget/render_object_to_widget_adaptor.dart';
-export 'src/widget/render_object_to_flutter_element_adapter.dart';
-export 'src/widget/character_data_to_flutter_element_adapter.dart';
-export 'src/widget/character_data_to_widget_adapter.dart';
-export 'src/widget/element_flutter_element_adapter.dart';
-export 'src/widget/element_to_widget_adapter.dart';
 export 'src/widget/widget_element.dart';
 export 'src/widget/webf_element.dart';
-export 'src/widget/webf_adapter_widget.dart';
 export 'src/widget/portal.dart';
 export 'src/widget/router_view.dart';
